+++ scripts/tmp-sweep.sh
@@ -510,6 +510,9 @@ cwdproj2-
 cwdproj3-
 cwdproj4-
 cwdreal-
+cx-commit-
+cx-commit-other-
+cx-commit-wt-
 env-scopes-
 envdef-home-
 envdef-home2-